Fixed BOTKUBE_VERSION (#483)
##### ISSUE TYPE
<!--- Pick one below and delete the rest: -->
 - Bug fix Pull Request

##### SUMMARY
<!--- Describe the change, including rationale and design decisions -->

The software notifies on the channel if you are using the updated version of BotKube or not. To do this, use the BOTKUBE_VERSION variable which is set up with the `.Chart.Version` value which does not necessarily correspond to the version of the software running on the cluster. The actual version of the software that runs is the one that comes from `.Values.image.tag`. 

<img width="616" alt="Screenshot 2021-05-11 at 09 51 39" src="https://user-images.githubusercontent.com/36044817/117780091-96588700-b23f-11eb-984c-b6555b1f3629.png">