feat(helm): allow to set deployment annotations (#537)
##### ISSUE TYPE

 - Feature Pull Request

##### SUMMARY


Allow to set deployment annotations in helm chart.

Fixes #536

##### Remarks

I hesitated to use `extraAnnotations`, but when I look at other official helm charts most of the time they use different values for pods and deployments annotations. 

Imho we should also rename `extraAnnotations` to `podAnnotations` but would be a breaking change so I left it as it was.